moved TextStyle + TextAlignment from draw.rs to text.rs
